masenet/requirements.txt
fairseq==0.12.2
kornia==0.8.0
matplotlib==3.9.4
plotly==6.2.0
numpy==1.26.4
optuna==4.4.0
optuna-dashboard==0.19.0
Pillow==11.2.1
pytorch_msssim==1.0.0
skimage==0.24.0
torch==2.7.0+cu128
torchvision==0.22.0+cu128
torchaudio==2.7.0+cu128
tqdm==4.65.2
rich==13.4.2
scipy==1.13.1
scikit-image==0.24.0
scikit-learn==1.6.1
pandas==2.3.1
imageio==2.37.0
pytorch-msssim==1.0.0
omegaconf==2.0.6
hydra-core==1.0.7
PyYAML==6.0.2
